drift-rl-agent/requirements.txt
tianshou>=1.0.0
gymnasium>=0.29.0
numpy>=1.24.0
requests>=2.31.0
openai>=1.0.0
pyyaml>=6.0
torch>=2.0.0
tensorboard>=2.14.0
